Inline and remove AttributesExtractor.
